[Feat] 현위치 주변 관광지 리스트 API, Recommendation 기본구조

+++ src/main/java/com/ssafy/enjoytrip/domain/attraction/mapper/AttractionMapper.java
@@ -56,8 +56,8 @@ public interface AttractionMapper {
 
     // AI 추천용 리스트 반환 (사용자 거리 + 반경 기반)
     List<AttractionForRecommendDTO> findAttractionsWithinRadiusForRecommend(
-            @Param("lat") double lat,
-            @Param("lng") double lng,
+            @Param("lat") BigDecimal lat,
+            @Param("lng") BigDecimal lng,
             @Param("radius") double radius
     );
 }
